Update EventHorizon.lua
Fixed event parsing in CLEU code, which solved:
- Ticks disappearing after crossing the "now" line
- Spell not reflecting haste changes

- however -

this also broke:
Ticks are delayed by a certain value (0.2 sec or something) when casting spells for the first time, behavior not observed before.
When recasting a spell ticks get assigned at wrong time, as if the spell was first cast instead of refreshing. this worked fine before.